test: stabilize Codex descendant cleanup coverage
